requirements.txt
streamlit
pdfplumber
PyPDF2
python-docx
pandas
pdf2image
pytesseract
pillow